% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAVplot.R
\name{ggAVplot}
\alias{ggAVplot}
\title{Generates a single Added Variable plot}
\usage{
ggAVplot(model, variable, data = NULL, ...)
}
\arguments{
\item{model}{An input model created by lm, glm, or any model that
supports `formula()` and `update()` functions.}

\item{variable}{The variable to create the Added Variable Plot on.}

\item{data}{A data frame that will be used for making the graph. This
is useful for adding labels, color encodings, etc for
other aspects of the graph the user wants to add.}

\item{...}{Further arguments to be passed to the whole plot `aes()`
function.}
}
\description{
Generates a single Added Variable plot
}
\examples{
model = lm( log(mpg) ~ disp + hp + drat + wt, data = mtcars)
ggAVplot(model, 'hp')
ggAVplot(model, 'hp', data=mtcars, color=cyl ) # add to plot-wide aesthetics list

}
